+++ locals.tf
@@ -16,4 +16,27 @@ locals {
       "S3_BUCKET_SERVER_SIDE_ENCRYPTION_ENABLED"
     ]
   )
+  monitor_iam_access = merge(
+    {
+      for identity in try(var.monitor_iam_access, []) : identity.name => {
+        "type"     = [identity.type]
+        "userName" = identity.name
+      } if identity.type == "IAMUser"
+    },
+    {
+      for identity in try(var.monitor_iam_access, []) : identity.name => {
+        "type" = [identity.type]
+        "sessionContext" = {
+          "sessionIssuer" = {
+            "userName" = identity.name
+          }
+        }
+      } if identity.type == "AssumedRole"
+    },
+    {
+      "Root" = {
+        "type" = ["Root"]
+      }
+    }
+  )
 }